# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gestion des containers Bridge Archipelago - création, arrêt, suppression et monitoring.",
        "title": "Archilan Orchestrateur API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sessions/{sessionId}/preflight": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates slot data (playerYaml, required options) and returns proposed slot names.\nStateless - does not create or modify any session record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Validate slots before generation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (informational only)",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Slot list",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PreflightRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PreflightResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "session not found"
                }
            }
        },
...
        "api.PreflightRequest": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.PreflightSlot"
                    }
                }
            }
        },
        "api.PreflightResponse": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.PreflightSlotResult"
                    }
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "api.PreflightSlot": {
            "type": "object",
            "properties": {
                "apworldStorageKey": {
                    "type": "string"
                },
                "archipelagoGameName": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.SlotOption"
                    }
                },
                "playerName": {
                    "type": "string"
                },
                "playerYaml": {
                    "type": "string"
                },
                "slotId": {
                    "type": "string"
                }
            }
        },
        "api.PreflightSlotResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "proposedName": {
                    "type": "string"
                },
                "slotId": {
                    "type": "string"
                }
            }
        },
...
        "api.SlotOption": {
            "type": "object",
            "properties": {
                "currentValue": {},
                "defaultValue": {},
                "key": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003cAPI_KEY\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}